Support older rubies
It looks like older versions of Ruby no longer work with newer versions
of OpenSSl.

In this case, ruby 2.7.6 would not install with openssl 1.1s.

To work around this we now compile OpenSSl 1.1.1g from source and pass
it to RUBY_CONFIGURE_OPTS when installing with rbenv.

It's not great but I think we are limited on options.